Use `source_files` in `test`
Instead of using `files` and specifying where to find the source
directory where these files live, use `source_files` added in
`conda-build` 2 to specify where these files live in the source
directory. This is done so that the files can be copied in for testing.
Has no other effect on the recipe or the packages built.